requirements.txt
accelerate==0.34.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
bitsandbytes==0.41.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.2
dill==0.3.7
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
# flash-attn==2.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
hf_transfer==0.1.8
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
latex2mathml==3.77.0
linkify-it-py==2.0.3
# -e git+https://github.com/seoulsky-field/xray-mllm-develop.git@20befd6fedad793647315e699fc8ba05b58a8f8f#egg=llava
markdown-it-py==2.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
narwhals==1.5.5
networkx==3.3
ninja==1.11.1.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.13.2
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==1.10.8
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.4
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.13.0
setproctitle==1.3.3
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.38.2
svgwrite==1.4.3
sympy==1.13.2
threadpoolctl==3.5.0
timm==1.0.7
tokenizers==0.21.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.47.0
triton==2.1.0
typing_extensions==4.12.2
tyro==0.8.8
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==1.26.19
uvicorn==0.30.6
wandb==0.17.7
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==13.0
xxhash==3.5.0
yarl==1.9.4
